Fixed duty cycle measurement code

+++ timer2.h
@@ -18,6 +18,6 @@
 // Function Prototypes
 // Configure for PWM Input Mode
 void timer2PWMInputModeInit(void);
-float normalizePWMTime(uint16_t duty_time);
+float normalizePWMTime(uint32_t duty_time);
 
 #endif
